added price errors

+++ src/main/resources/sql-scripts/CreateTables.sql
@@ -121,7 +121,7 @@ CREATE TABLE review (
     prod_id VARCHAR(255) REFERENCES product (prod_id) ON DELETE CASCADE,
     username VARCHAR(50) NOT NULL,
     reviewdate DATE NOT NULL,
-    rating INT NOT NULL,
+    rating FLOAT NOT NULL,
     helpful_rating INT NOT NULL,
     review_sum VARCHAR(255),
     review_text VARCHAR(5000),
